refactor(hooks): generify agent installation and move Claude-specific config into ClaudeAdapter
